升级SpringBoot 2.7.0，解决循环依赖问题
