+++ src/main/java/com/macro/mall/tiny/modules/ums/service/UmsAdminService.java
@@ -81,4 +81,9 @@ public interface UmsAdminService extends IService<UmsAdmin> {
      * 获取用户信息
      */
     UserDetails loadUserByUsername(String username);
+
+    /**
+     * 获取缓存服务
+     */
+    UmsAdminCacheService getCacheService();
 }
