+++ src/main/java/com/macro/mall/tiny/security/component/DynamicSecurityFilter.java
@@ -20,8 +20,6 @@ import java.io.IOException;
  * 动态权限过滤器，用于实现基于路径的动态权限过滤
  * Created by macro on 2020/2/7.
  */
-@Component
-@ConditionalOnBean(name = "dynamicSecurityService")
 public class DynamicSecurityFilter extends AbstractSecurityInterceptor implements Filter {
 
     @Autowired
