+++ src/main/java/com/macro/mall/tiny/security/component/JwtAuthenticationTokenFilter.java
@@ -23,7 +23,6 @@ import java.io.IOException;
  * JWT登录授权过滤器
  * Created by macro on 2018/4/26.
  */
-@Component
 public class JwtAuthenticationTokenFilter extends OncePerRequestFilter {
     private static final Logger LOGGER = LoggerFactory.getLogger(JwtAuthenticationTokenFilter.class);
     @Autowired
